+++ doc/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 1.1
+
+- ADD: implement a speculative policy in query (~200% speedup)
+- ADD: implement a simple block decomposition algorithm `rmq_block` (thanks to Songyang Chen)
+
 ## Version 1.0
 
-- ADD: Implement naive, sparse table and indirection solver.
+- ADD: implement naive, sparse table and indirection solver